udpdated type definition files and consolidated types where possible
