+++ client/components/Dashboard.tsx
@@ -5,8 +5,6 @@ import { useQuery, gql } from '@apollo/client';
 import ToolBar from './ToolBar';
 import ProjectView from './ProjectView';
 import { useAuth } from '../auth/AuthProvider';
-import { RateLimiterConfig } from '../../@types/dashboard';
-// import { SelectedProject, Projects } from '../../@types/Interfaces';
 
 const GET_PROJECT_DATA = gql`
     query getUserData($userId: String!) {
