+++ package.json
@@ -65,6 +65,7 @@
         "@types/graphql": "^14.5.0",
         "@types/jsonwebtoken": "^8.5.8",
         "@types/lodash": "^4.14.182",
+        "@types/mongoose": "^5.11.97",
         "@types/node": "^17.0.33",
         "@types/randomstring": "^1.1.8",
         "@types/react": "^18.0.9",
