+++ tsconfig.json
@@ -1,7 +1,11 @@
 {
     "compilerOptions": {
         "target": "es5",
-        "lib": ["dom", "dom.iterable", "esnext"],
+        "lib": [
+            "dom",
+            "dom.iterable",
+            "esnext"
+        ],
         "allowJs": true,
         "skipLibCheck": true,
         "esModuleInterop": true,
@@ -15,7 +19,10 @@
         "isolatedModules": true,
         "noEmit": false,
         "jsx": "react-jsx",
-        "typeRoots": ["./@types", "node_modules/@types"]
+        "typeRoots": [
+            "./@types",
+            "node_modules/@types"
+        ]
     },
     "include": [
         "client/**/*.js",
